flightcomp/requirements.txt
pillow==10.1.0
pyttsx3==2.90
pygame==2.5.2 